fix(tests): pytest collection fix — remove src.* mocking + INT8 dequant dtype cast
- INT8KVCache.get_kv() now casts dequant output to self.dtype (fp16)
- conftest.py: removed SRC_DIR from sys.path (caused package conflict)
- test_eval_ppl_guardrails.py: stopped mocking src.* packages
- Added tearDownModule to 3 files with sys.modules mocking
Result: 737 passed (up from 0), 290 failed (test assertions need sync)

+++ src/cache/int8_cache.py
@@ -534,11 +534,12 @@ class INT8KVCache:
         q_v = self._v_cache[layer_id][:, :, :seq_len, :]
         scale_v = self._v_scale[layer_id][:, :, :seq_len, :]
 
-        # Dequantize
+        # Dequantize (scale is fp32, dequant returns fp32)
         k = dequantize_symmetric_int8(q_k, scale_k)
         v = dequantize_symmetric_int8(q_v, scale_v)
 
-        return k, v
+        # Cast back to model dtype (fp16) for API compatibility
+        return k.to(self.dtype), v.to(self.dtype)
 
     def get_int8_tensors(self, layer_id: int) -> Tuple[Tensor, Tensor, Tensor, Tensor]:
         """
